requirements.txt
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
